Remove the old FlexmiYamlParser
